renamed requests to requests_auth
